requirement.txt
pandas
numpy
scikit-learn
boto3
python-dotenv
PyYAML
dataclasses